Skeleton of the dynamic expression system
