+++ include/xtensor/zfunction.hpp
@@ -1,0 +1,18 @@
+/***************************************************************************
+* Copyright (c) Johan Mabille, Sylvain Corlay and Wolf Vollprecht          *
+* Copyright (c) QuantStack                                                 *
+*                                                                          *
+* Distributed under the terms of the BSD 3-Clause License.                 *
+*                                                                          *
+* The full license is in the file LICENSE, distributed with this software. *
+****************************************************************************/
+
+#ifndef XTENSOR_ZFUNCTION_HPP
+#define XTENSOR_ZFUNCTION_HPP
+
+namespace xt
+{
+}
+
+#endif
+
